Set up gradle project
