+++ bindings/java/src/main/java/org/github/tursodatabase/limbo/Connection.java
@@ -1,4 +1,4 @@
-package limbo;
+package org.github.tursodatabase.limbo;
 
 import java.lang.Exception;
